+++ lib/DocblockParser/Lexer.php
@@ -25,7 +25,7 @@ final class Lexer
         '=', // equals
         '(', ')', '\{', '\}', '\[', '\]', '<', '>', // brackets
         '\$[a-zA-Z0-9_\x80-\xff]+', // variable
-        '[a-zA-Z0-9_\\\]+', // label
+        '[a-zA-Z\\\][-a-zA-Z0-9_\\\]*', // label
     ];
     private const TOKEN_VALUE_MAP = [
         ']' => Token::T_BRACKET_SQUARE_CLOSE,